Fixes tests & refactors index.js
