+++ lib/util.js
@@ -117,6 +117,6 @@ function distanceConvert (value, unit) {
  */
 module.exports.firstToLowerCase = firstToLowerCase;
 function firstToLowerCase(string) {
-    var modified = string.substr(0, 1).toLowerCase() + string.substr(1);
-    return modified;
-};
+  var modified = string.substr(0, 1).toLowerCase() + string.substr(1);
+  return modified;
+}
